requirements.txt
fastapi
uvicorn
sarvamai
pydub
google-generativeai
